Move autoscroll enum to bulletproof structure
