+++ src/apps/stable/features/lyrics/constants/autoScroll.ts
